starting to do more sophisticated styling
