+++ GeoChat/styles/homepage.js
@@ -1,0 +1,28 @@
+
+import { StyleSheet } from 'react-native';
+const styles = StyleSheet.create({
+    container: {
+        flex: 1,
+        backgroundColor: '#fff',
+        alignItems: 'center',
+        justifyContent: 'center',
+    },
+    header : {
+        fontSize : 30
+    },
+    subText : {
+        fontSize : 12,
+        width : 200,
+        textAlign : "center",
+ 
+    },
+
+  red: {
+    color: 'red',
+  }
+
+
+});
+
+
+export default styles;
